requirements.txt
captcha==0.3
gevent==1.4.0
greenlet==0.4.15
celery==4.4.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.11.1
Pillow==7.0.0
PyMySQL==0.9.3
python-alipay-sdk==2.0.1
qiniu==7.2.6
redis==3.4.1
requests==2.23.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==1.0.0
WTForms==2.2.1
